pyyaml and toml are optional dependencies.
Users are likely to only need one and not the other.

It seems heavy handed to require both, or to expand this out to a package which has permutations of these as dependencies.
I think it'd be better to not have either as dependencies and let end users install any parsers they need.